Update User.cpp
Switched over to using vector based lists to track liked people, films, genres.
Added addLikedFilm method, which adds films to the vector lists, as well as the people in the Film, and the genre of the Film.
Added scoreFilm(Film f) method, which iterates over the vector lists of a user, and compares them to a given film. return an int value, the higher , the more highly recommended.